% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDMA.R
\name{wrapCATE}
\alias{wrapCATE}
\title{wrapCATE : Mediation EWAS and independent mediation analysis (coufounders estimated via CATE method)}
\usage{
wrapCATE(X, Y, M, K, conf.known = NULL, mediation = T, top = 50,
  sims = 100, ...)
}
\arguments{
\item{X}{exposure}

\item{Y}{outcome}

\item{M}{methylation matrix}

\item{K}{number of coufounders to estimate}

\item{conf.known}{known confondeurs}

\item{mediation}{TRUE or FALSE, if TRUE : independant mediation analysis}

\item{top}{If mediation is true, then "top" is the number of CpGs analyzed by wrapMed()}

\item{sims}{Number of MC simulations in the mediation model}

\item{...}{other options of the CATE model}
}
\value{
pval, qval, score for each CpG (note: the pValues are calibrated)

conf : coufounders estimated via CATE method

media : if mediation = T, result of wrapMed() function

mod : whole CATE model
}
\description{
wrapCATE : Mediation EWAS and independent mediation analysis (coufounders estimated via CATE method)
}
\details{
This function estimates the confusers via the CATE model. 
In addition, the CATE model uses robust linear regressions "rlm" to perform EWAS associations testing of mediation.
}
\examples{
# Simulate data :
simu <- r_mediation(100, 500, 5)
# do a mediation EWAS and the independant mediation analysis
mod <- wrapCATE(simu$X, simu$Y, 5, mediation = T)
# quick result of mediation EWAS
plot(-log10(mod$pval))
points(simu$mediators, -log10(mod$pval[simu$mediators]),col = 2)
}
