% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDMA.R
\name{wrapMed}
\alias{wrapMed}
\title{wrapper of mediation package (for one mediator)}
\usage{
wrapMed(X, Y, mediator, CONF = NULL, sims = 100, ...)
}
\arguments{
\item{X}{exposure (continuous)}

\item{Y}{outcome (continuous or factor)}

\item{mediator}{one CpG site}

\item{CONF}{estimated and known confounders}

\item{sims}{number of simulations}

\item{...}{other options for the mediation model}
}
\value{
a pValue for ACME (ACME.p.med)

a effect for ACME (ACME.eff)
}
\description{
wrapper of mediation package (for one mediator)
}
\details{
This function makes it possible to carry out an independent mediation analysis between an exposure (X),
 a CpG (M) and an outcome (Y). For continuous X and Y it carries out a conbinaison of 2 linear models.
  For a continuous X and a factorial Y it is a conbinaison of a linear model and a GLM.
  With these conbinations, it calculates an ACME (Average Causal Mediation Effects) and an associated value.
}
\examples{
# Simulate data :
simu <- r_mediation(100, 500, 5)
# test if the first CpG mediate a effect
mod <- wrapMed(simu$X, simu$Y, simu$M[,1])
}
