% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDMA.R
\name{wrapBACO}
\alias{wrapBACO}
\title{wrapBACO : Mediation EWAS and independent mediation analysis (no coufounders estimated)}
\usage{
wrapBACO(X, Y, M, conf.known = NULL, mediation = T, top = 50,
  GLMwrap = NULL, sims = 100)
}
\arguments{
\item{X}{exposure}

\item{Y}{outcome}

\item{M}{methylation matrix}

\item{conf.known}{known confondeurs}

\item{mediation}{TRUE or FALSE, if TRUE : independant mediation analysis}

\item{top}{If mediation is true, then "top" is the number of CpGs analyzed by wrapMed()}

\item{GLMwrap}{result of wrapGLM}

\item{sims}{Number of MC simulations in the mediation model}
}
\value{
pval, qval, score for each CpG

media : if mediation = T, result of wrapMed() function

mod : whole bacon model
}
\description{
wrapBACO : Mediation EWAS and independent mediation analysis (no coufounders estimated)
}
\details{
The BACON model does not consider confusers but it aims to reduce the bias and inflation
 of a statistical test that in our case is a glm. 
Note that this method can be combined with a method estimating confondeurs.
}
\examples{
# Simulate data :
simu <- r_mediation(100, 500, 5)
# do a mediation EWAS and the independant mediation analysis
mod <- wrapBACO(simu$X, simu$Y, mediation = T)
# quick result of mediation EWAS
plot(-log10(mod$pval))
points(simu$mediators, -log10(mod$pval[simu$mediators]),col = 2)

# or with a previous glm
mod1 <- wrapGLM(simu$X, simu$Y, mediation = F)
mod2 <- wrapBACO(simu$X, simu$Y, mediation = T, GLMwrap = mod1)
}
