% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDMA.R
\name{wrapLFMM}
\alias{wrapLFMM}
\title{wrapLFMM : Mediation EWAS and independent mediation analysis (coufounders estimated via LFMM methods)}
\usage{
wrapLFMM(X, Y, M, K, meth = c("ridge", "lasso"), conf.known = NULL,
  mediation = T, top = 50, sims = 100, ...)
}
\arguments{
\item{X}{exposure}

\item{Y}{outcome}

\item{M}{methylation matrix}

\item{K}{number of coufounders to estimate}

\item{meth}{LFMM combines two methods to estimate coufondeurs: "ridge" and "lasso"}

\item{conf.known}{known confondeurs}

\item{mediation}{TRUE or FALSE, if TRUE : independant mediation analysis}

\item{top}{If mediation is true, then "top" is the number of CpGs analyzed by wrapMed()}

\item{sims}{Number of MC simulations in the mediation model}

\item{...}{other options of the LFMMs models}
}
\value{
pval, qval, score for each CpG (note: the pValues are calibrated)

conf : coufounders estimated via LFMM methods

media : if mediation = T, result of wrapMed() function

mod : whole LFMM model
}
\description{
wrapLFMM : Mediation EWAS and independent mediation analysis (coufounders estimated via LFMM methods)
}
\details{
This function estimates the confusers via the LFMM model. There are two sub-models of LFMM to estimate the confondeurs, 
"lasso" and "ridge".  
In addition, the LFMM model uses linear regression "lm" to perform associations tests of mediation EWAS.
}
\examples{
# Simulate data :
simu <- r_mediation(100, 500, 5)
# do a mediation EWAS and the independant mediation analysis (LFMM ridge)
mod <- wrapLFMM(simu$X, simu$Y, 5, mediation = T,meth = "ridge")
# quick result of mediation EWAS
plot(-log10(mod$pval))
points(simu$mediators, -log10(mod$pval[simu$mediators]),col = 2)
}
