% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDMA.R
\name{wrapRFE}
\alias{wrapRFE}
\title{wrapRFE : multivarite ewas and mediation function (coufounders estimated via RefFreeEWAS methods)}
\usage{
wrapRFE(X, Y, M, K, nbboot = 50, conf.known = NULL, mediation = T,
  top = 50, sims = 100, ...)
}
\arguments{
\item{X}{exposure}

\item{Y}{outcome}

\item{M}{methylation matrix}

\item{K}{number of coufounders to estimate}

\item{nbboot}{number of bootstrap to calculate pValue}

\item{conf.known}{known confondeurs}

\item{mediation}{TRUE or FALSE, if TRUE : independant mediation analysis}

\item{top}{If mediation is true, then "top" is the number of CpGs analyzed by wrapMed()}

\item{sims}{Number of MC simulations in the mediation model}

\item{...}{other options of the RefFreeEWAS model}
}
\value{
pval, qval, score for each CpG

conf : coufounders estimated via RefFreeEWAS methods

media : if mediation = T, result of wrapMed() function

mod : whole RefFreeEWAS model
}
\description{
wrapRFE : multivarite ewas and mediation function (coufounders estimated via RefFreeEWAS methods)
}
\examples{
# Simulate data :
simu <- r_mediation(100, 500, 5)
# do a multivariate ewas and the independant mediation analysis
mod <- wrapRFE(simu$X, simu$Y, 5, mediation = T)
# quick result of multivariate ewas
plot(-log10(mod$pval))
points(simu$mediators, -log10(mod$pval[simu$mediators]),col = 2)
}
