% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDMA.R
\name{wrapSVA}
\alias{wrapSVA}
\title{wrapSVA : multivarite ewas and mediation function (coufounders estimated via SVA method)}
\usage{
wrapSVA(X, Y, M, K, controls = NULL, conf.known = NULL, mediation = T,
  top = 50, sims = 100, ...)
}
\arguments{
\item{X}{exposure}

\item{Y}{outcome}

\item{M}{methylation matrix}

\item{K}{number of coufounders to estimate}

\item{controls}{NOT USE}

\item{conf.known}{known confondeurs}

\item{mediation}{TRUE or FALSE, if TRUE : independant mediation analysis}

\item{top}{If mediation is true, then "top" is the number of CpGs analyzed by wrapMed()}

\item{sims}{Number of MC simulations in the mediation model}

\item{...}{other options of the SVA model}
}
\value{
pval, qval, score for each CpG

conf : coufounders estimated via SVA method

media : if mediation = T, result of wrapMed() function

mod : whole SVA model
}
\description{
wrapSVA : multivarite ewas and mediation function (coufounders estimated via SVA method)
}
\examples{
# Simulate data :
simu <- r_mediation(100, 500, 5)
# do a multivariate ewas and the independant mediation analysis
mod <- wrapSVA(simu$X, simu$Y, 5, mediation = T)
# quick result of multivariate ewas
plot(-log10(mod$pval))
points(simu$mediators, -log10(mod$pval[simu$mediators]),col = 2)
}
